Do not error out if a directory already exists.
Upstream seem to assume that each node is using its own private
directory as the tmpdir base, but we have no local disk and a
shared fast scratch filesystem instead.

So do not treat EEXIST as an error as for multi-node jobs this
be attempting to create the directory in multiple places otherwise
nodes will fail with "Prolog error".